requirements/production.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=/app/requirements/production.txt /app/requirements/in/base.in /app/requirements/in/production.in
#
--no-binary psycopg2

amqp==5.0.6
    # via kombu
argon2-cffi==20.1.0
    # via -r /app/requirements/in/base.in
asgiref==3.3.4
    # via django
azure-common==1.1.27
    # via
    #   azure-mgmt-cdn
    #   azure-storage-blob
    #   azure-storage-common
azure-core==1.13.0
    # via
    #   azure-identity
    #   azure-mgmt-core
azure-identity==1.5.0
    # via -r /app/requirements/in/base.in
azure-mgmt-cdn==11.0.0
    # via -r /app/requirements/in/base.in
azure-mgmt-core==1.2.2
    # via azure-mgmt-cdn
azure-storage-blob==2.1.0
    # via django-storages
azure-storage-common==2.1.0
    # via azure-storage-blob
backcall==0.2.0
    # via ipython
beautifulsoup4==4.9.3
    # via -r /app/requirements/in/base.in
billiard==3.6.4.0
    # via celery
cachetools==4.2.2
    # via steam
celery==5.0.5
    # via
    #   -r /app/requirements/in/base.in
    #   django-celery-beat
    #   django-celery-results
certifi==2020.12.5
    # via
    #   httpx
    #   msrest
    #   requests
cffi==1.14.5
    # via
    #   argon2-cffi
    #   cryptography
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   -r /app/requirements/in/base.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   django-click
colorama==0.4.4
    # via -r /app/requirements/in/base.in
coloredlogs==15.0
    # via -r /app/requirements/in/base.in
cryptography==3.4.7
    # via
    #   azure-identity
    #   azure-storage-common
    #   msal
    #   pyjwt
dateparser==1.0.0
    # via -r /app/requirements/in/base.in
decorator==5.0.7
    # via ipython
django-celery-beat==2.2.0
    # via -r /app/requirements/in/base.in
django-celery-results==2.0.1
    # via -r /app/requirements/in/base.in
django-click==2.2.0
    # via -r /app/requirements/in/base.in
django-cors-headers==3.7.0
    # via -r /app/requirements/in/base.in
django-crispy-forms==1.11.2
    # via -r /app/requirements/in/base.in
django-environ==0.4.5
    # via -r /app/requirements/in/base.in
django-filter==2.4.0
    # via -r /app/requirements/in/base.in
django-json-widget==1.1.1
    # via -r /app/requirements/in/base.in
django-model-utils==4.1.1
    # via -r /app/requirements/in/base.in
django-polymorphic==3.0.0
    # via -r /app/requirements/in/base.in
django-prometheus==2.1.0
    # via -r /app/requirements/in/base.in
django-redis==4.12.1
    # via
    #   -r /app/requirements/in/base.in
    #   python-redis-lock
django-rest-fuzzysearch==0.5.1
    # via -r /app/requirements/in/base.in
django-robots==4.0
    # via -r /app/requirements/in/base.in
django-storages[azure]==1.11.1
    # via -r /app/requirements/in/base.in
django-timezone-field==4.1.2
    # via django-celery-beat
django==3.2
    # via
    #   -r /app/requirements/in/base.in
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-model-utils
    #   django-polymorphic
    #   django-redis
    #   django-rest-fuzzysearch
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-msgpack
git+https://github.com/AngellusMortis/django-rest-framework-msgpack.git
    # via -r /app/requirements/in/base.in
djangorestframework==3.12.4
    # via
    #   -r /app/requirements/in/base.in
    #   django-rest-fuzzysearch
    #   djangorestframework-msgpack
    #   drf-extensions
drf-extensions==0.7.0
    # via -r /app/requirements/in/base.in
et-xmlfile==1.1.0
    # via openpyxl
filetype==1.0.7
    # via -r /app/requirements/in/base.in
future==0.18.2
    # via django-json-widget
gevent-eventemitter==2.1
    # via steam
gevent==21.1.2
    # via
    #   gevent-eventemitter
    #   steam
greenlet==1.0.0
    # via gevent
h11==0.12.0
    # via httpcore
h2==3.2.0
    # via httpx
hpack==3.0.0
    # via h2
httpcore==0.13.2
    # via httpx
httpx[http2]==0.18.1
    # via -r /app/requirements/in/base.in
huey==2.3.2
    # via -r /app/requirements/in/base.in
humanfriendly==9.1
    # via coloredlogs
hyperframe==5.2.0
    # via h2
idna==2.10
    # via
    #   requests
    #   rfc3986
ipython-genutils==0.2.0
    # via traitlets
ipython==7.23.0
    # via -r /app/requirements/in/base.in
isodate==0.6.0
    # via msrest
jedi==0.18.0
    # via ipython
kombu==5.0.2
    # via celery
matplotlib-inline==0.1.2
    # via ipython
msal-extensions==0.3.0
    # via azure-identity
msal==1.11.0
    # via
    #   azure-identity
    #   msal-extensions
msgpack==1.0.2
    # via
    #   -r /app/requirements/in/base.in
    #   djangorestframework-msgpack
msrest==0.6.21
    # via azure-mgmt-cdn
numpy==1.20.2
    # via -r /app/requirements/in/base.in
oauthlib==3.1.0
    # via requests-oauthlib
openpyxl==3.0.7
    # via -r /app/requirements/in/base.in
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.2.0
    # via -r /app/requirements/in/base.in
portalocker==1.7.1
    # via msal-extensions
prometheus-client==0.10.1
    # via django-prometheus
prompt-toolkit==3.0.18
    # via
    #   click-repl
    #   ipython
protobuf==3.15.8
    # via steam
psycopg2==2.8.6
    # via -r /app/requirements/in/base.in
ptyprocess==0.7.0
    # via pexpect
pycparser==2.20
    # via cffi
pycryptodomex==3.10.1
    # via steam
pydiscourse==1.1.2
    # via -r /app/requirements/in/base.in
pygments==2.8.1
    # via ipython
pyjwt[crypto]==2.1.0
    # via msal
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.8.1
    # via
    #   azure-storage-common
    #   dateparser
    #   djangorestframework-msgpack
    #   python-crontab
python-redis-lock[django]==3.7.0
    # via -r /app/requirements/in/base.in
python-slugify==5.0.0
    # via -r /app/requirements/in/base.in
pytz==2021.1
    # via
    #   celery
    #   dateparser
    #   django
    #   django-timezone-field
    #   tzlocal
pyyaml==5.4.1
    # via -r /app/requirements/in/base.in
redis==3.5.3
    # via
    #   django-redis
    #   python-redis-lock
regex==2021.4.4
    # via dateparser
requests-oauthlib==1.3.0
    # via msrest
requests==2.25.1
    # via
    #   azure-core
    #   azure-storage-common
    #   msal
    #   msrest
    #   pydiscourse
    #   requests-oauthlib
    #   steam
rfc3986[idna2008]==1.4.0
    # via httpx
semver==2.13.0
    # via -r /app/requirements/in/base.in
six==1.15.0
    # via
    #   argon2-cffi
    #   azure-core
    #   azure-identity
    #   click-repl
    #   isodate
    #   protobuf
    #   python-dateutil
    #   steam
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via django
steam[client]==1.2.0
    # via -r /app/requirements/in/base.in
text-unidecode==1.3
    # via python-slugify
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
tzlocal==2.1
    # via dateparser
uritemplate==3.0.1
    # via -r /app/requirements/in/base.in
urllib3==1.26.4
    # via requests
uwsgi==2.0.19.1
    # via -r /app/requirements/in/production.in
vdf==3.3
    # via steam
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
zope.event==4.5.0
    # via gevent
zope.interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
